Multi-Window Issues
Attempt to allow multiple windows to be spawned with the class. 

In testing, runs into a RuntimeError: main thread is not in main loop error if attempting to display second window after closing first despite already quitting the mainloop.  Even if this were fixed, trying to display both windows at once would also cause this error because it would attempt to start a second mainloop.